StEP00204: Dateien eines Nutzers nicht löschen, wenn der Nutzer gelöscht wird re #1293
git-svn-id: svn://develop.studip.de/studip/trunk@16420 c5b85792-5220-0410-ad23-ed5b6b8ad999

+++ lib/classes/UserManagement.class.php
@@ -642,9 +642,10 @@ class UserManagement
     * Delete an existing user from the database and tidy up
     *
     * @access   public
+    * @param    bool delete all documents belonging to the user
     * @return   bool Removal successful?
     */
-    function deleteUser() {
+    function deleteUser($delete_documents = true) {
         global $perm, $auth;
 
         // Do we have permission to do so?
@@ -725,43 +726,44 @@ class UserManagement
         $user_language = getUserLanguagePath($this->user_data['auth_user_md5.user_id']);
 
         // delete documents of this user
-        $temp_count = 0;
-        $query = "SELECT dokument_id FROM dokumente WHERE user_id='" . $this->user_data['auth_user_md5.user_id'] . "'";
-        $this->db->query($query);
-        while ($this->db->next_record()) {
-            if (delete_document($this->db->f("dokument_id")))
-                $temp_count ++;
+        if ($delete_documents) {
+	        $temp_count = 0;
+	        $query = "SELECT dokument_id FROM dokumente WHERE user_id='" . $this->user_data['auth_user_md5.user_id'] . "'";
+	        $this->db->query($query);
+	        while ($this->db->next_record()) {
+	            if (delete_document($this->db->f("dokument_id")))
+	                $temp_count ++;
+	        }
+	        if ($temp_count) {
+	            $this->msg .= "info§" . sprintf(_("%s Dokumente gel&ouml;scht."), $temp_count) . "§";
+	        }
+	
+	        // delete empty folders of this user
+	        $temp_count = 0;
+	        $query = "SELECT folder_id FROM folder WHERE user_id='" . $this->user_data['auth_user_md5.user_id'] . "' ORDER BY mkdate DESC";
+	        $this->db->query($query);
+	        while ($this->db->next_record()) {
+	            $query = "SELECT count(*) AS count FROM folder WHERE range_id = '".$this->db->f("folder_id")."'";
+	            $this->db2->query($query);
+	            $this->db2->next_record();
+	            if (!$this->db2->f("count") && !doc_count($this->db->f("folder_id"))) {
+	                $query = "DELETE FROM folder WHERE folder_id ='".$this->db->f("folder_id")."'";
+	                $this->db2->query($query);
+	                $temp_count += $this->db2->affected_rows();
+	            }
+	        }
+	        if ($temp_count) {
+	            $this->msg .= "info§" . sprintf(_("%s leere Ordner gel&ouml;scht."), $temp_count) . "§";
+	        }
+	
+	        // folder left?
+	        $query = "SELECT count(*) AS count FROM folder WHERE user_id='" . $this->user_data['auth_user_md5.user_id'] . "'";
+	        $this->db->query($query);
+	        $this->db->next_record();
+	        if ($this->db->f("count")) {
+	            $this->msg .= sprintf("info§" . _("%s Ordner konnten nicht gel&ouml;scht werden, da sie noch Dokumente anderer BenutzerInnen enthalten.") . "§", $this->db->f("count"));
+	        }
         }
-        if ($temp_count) {
-            $this->msg .= "info§" . sprintf(_("%s Dokumente gel&ouml;scht."), $temp_count) . "§";
-        }
-
-        // delete empty folders of this user
-        $temp_count = 0;
-        $query = "SELECT folder_id FROM folder WHERE user_id='" . $this->user_data['auth_user_md5.user_id'] . "' ORDER BY mkdate DESC";
-        $this->db->query($query);
-        while ($this->db->next_record()) {
-            $query = "SELECT count(*) AS count FROM folder WHERE range_id = '".$this->db->f("folder_id")."'";
-            $this->db2->query($query);
-            $this->db2->next_record();
-            if (!$this->db2->f("count") && !doc_count($this->db->f("folder_id"))) {
-                $query = "DELETE FROM folder WHERE folder_id ='".$this->db->f("folder_id")."'";
-                $this->db2->query($query);
-                $temp_count += $this->db2->affected_rows();
-            }
-        }
-        if ($temp_count) {
-            $this->msg .= "info§" . sprintf(_("%s leere Ordner gel&ouml;scht."), $temp_count) . "§";
-        }
-
-        // folder left?
-        $query = "SELECT count(*) AS count FROM folder WHERE user_id='" . $this->user_data['auth_user_md5.user_id'] . "'";
-        $this->db->query($query);
-        $this->db->next_record();
-        if ($this->db->f("count")) {
-            $this->msg .= sprintf("info§" . _("%s Ordner konnten nicht gel&ouml;scht werden, da sie noch Dokumente anderer BenutzerInnen enthalten.") . "§", $this->db->f("count"));
-        }
-
         // kill all the ressources that are assigned to the user (and all the linked or subordinated stuff!)
         if ($GLOBALS['RESOURCES_ENABLE']) {
             $killAssign = new DeleteResourcesUser($this->user_data['auth_user_md5.user_id']);
